API Calls for Excercises Screen 2
